requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
llama-cpp-python==0.3.9
requests==2.31.0
pydantic==2.11.4
python-dotenv==1.0.1 